Hourly Commit 7? (Day 9 or 10)
I've focused on the devlogs so much I forgot to actually commit! Anyways, I changed the paddle hitbox back to a rectangle, gave it a unique collision layer, changed the color of the bricks, and modified the bouncing when the ball hits the paddle. It's a little janky for now, but rest assured I'm working on it!