Hourly Commit 8.5 (Day 10)
Removed the paddle bouncing thing until I can figure it out later, and changed the hitbox back to capsule. Also, forgot to mention last commit, but I added a life system as well.

In this commit, I added a score system, organized the code a little bit, and added screen shake for when you hit a brick or lose a life. Also added a flashing effect for when you lose a life that appears on the life visual.